Support selecting GDScript-style identifiers.
There are three changes in this patch, to support double-tapping on a
word to extend the selection when there are underscore letters (which
is what GDScript supports).

First, rather than using `CharacterSet.alphanumerics` as the filter, I
add the underscore to the set that is used to check for validity.

Second, rather that continuing a scan when the initial character being
probbed is a valid identifier set, we stop rigth there.  Otherwise
double-tapping on the word 'world' in "hello & world" would select
also the ampersand, like this: "& world".

Lastly, this implements the rangeEnclosingPosition(:with:inDirection)
method, which given a position scans backwards and foreward with the
new rules to find a match in `.word` mode.